+++ flutter_audio_games/lib/flutter_audio_games.dart
@@ -28,6 +28,7 @@ export 'src/sounds/sound_type.dart';
 export 'src/sounds/source_loader.dart';
 export 'src/sounds/voice_group.dart';
 export 'src/turning_direction.dart';
+export 'src/widgets//sounds/protect_sounds.dart';
 export 'src/widgets/audio_game_menu/audio_game_menu.dart';
 export 'src/widgets/audio_game_menu/audio_game_menu_item.dart';
 export 'src/widgets/audio_game_menu/audio_game_menu_item_list_tile.dart';